requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Vector Store
pinecone-client==3.0.0

# Embeddings & LLM
openai==1.10.0
tiktoken==0.5.2

# CrewAI
crewai==0.1.26
crewai-tools==0.0.14
langchain==0.1.0
langchain-openai==0.0.2

# OAuth & Security
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.115.0
msal==1.26.0
cryptography==42.0.0

# Email Processing
email-validator==2.1.0.post1
beautifulsoup4==4.12.3
html2text==2020.1.16

# Background Jobs
apscheduler==3.10.4
celery==5.3.6
redis==5.0.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3

# Monitoring & Logging
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.6.0

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0
pre-commit==3.6.0
